Added a "--notest" to the compile for release script.  This is helpful for interating on a release.

+++ compile-for-publish.sh
@@ -10,9 +10,17 @@ if [ ! -f ./package.json ]; then
 	echo Please run this from the repo root folder
 	exit 1
 fi
+
+if [ "$1" != "" ] && [ "$1" != "--notest" ]; then
+  echo "ERROR: $1 is not a recognized option"
+  exit 1
+fi
+
 npm install
 node node_modules/streamline/bin/_node -c lib
 find lib/ -name "*._js" -delete
 node bin/azure --gen
-node scripts/unit.js testlist.txt
-node scripts/unit.js testlist-arm.txt
+if [ "$1" != "--notest" ]; then
+  node scripts/unit.js testlist.txt
+  node scripts/unit.js testlist-arm.txt
+fi
